requirements_web.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
requests>=2.31.0
langchain==0.1.4
langchain-community>=0.0.14
sentence-transformers==2.5.0
chromadb==0.4.22
openai==0.28.0 